lunchmate/requirements.txt
asgiref==3.8.1
astroid==3.2.4
black==24.8.0
click==8.1.7
dill==0.3.8
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
PyJWT==2.9.0
pylint==3.2.7
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
